backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
aiomysql==0.2.0
cryptography
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiohttp==3.10.0
apscheduler==3.10.4
pydantic==2.9.0
pydantic-settings==2.5.2
